Cancel script execution at failures but exit with 0
Only the last build stage can exit with non-zero code. This way
hsl, waltti and finland data images get build independently although
any of builds fail.

+++ build-docker-image.sh
@@ -22,21 +22,42 @@ echo "***Launching the builder"
 rm -rf target build
 mkdir target
 docker run --rm --entrypoint tar $DOCKER_BUILDER_IMAGE -c /opt/$CONTAINER/webroot|tar x -C target
+if [ $? -ne 0 ]; then
+    exit 0
+fi
 echo "*** Builder launched"
 
 #prebuild graph
 echo "*** Prebuilding the graph"
 mkdir -p build/$ROUTER_NAME
 unzip -j target/opt/$CONTAINER/webroot/router-$ROUTER_NAME.zip -d build/$ROUTER_NAME/
+if [ $? -ne 0 ]; then
+    exit 0
+fi
 VERSION=`docker run --rm --entrypoint /bin/bash hsldevcom/opentripplanner:prod  -c "java -jar otp-shaded.jar --version"|grep commit|cut -d' ' -f2`
 docker run -v `pwd`/build:/opt/opentripplanner/graphs --rm --entrypoint /bin/bash hsldevcom/opentripplanner:prod  -c "java -Xmx6g -jar otp-shaded.jar --build graphs/$ROUTER_NAME"
+if [ $? -ne 0 ]; then
+    exit 0
+fi
 cd build/
 zip graph-$ROUTER_NAME-$VERSION.zip $ROUTER_NAME/router-config.json $ROUTER_NAME/Graph.obj
+if [ $? -ne 0 ]; then
+    exit 0
+fi
 cd ..
 
 #build docker image
 echo "*** Building the docker image"
 docker build --tag=$DOCKER_TAGGED_IMAGE -f Dockerfile .
+if [ $? -ne 0 ]; then
+    exit 0
+fi
 docker login -u $DOCKER_USER -p $DOCKER_AUTH
+if [ $? -ne 0 ]; then
+    exit 0
+fi
 docker push $DOCKER_TAGGED_IMAGE
+if [ $? -ne 0 ]; then
+    exit 0
+fi
 echo "*** $ROUTER_NAME image built and pushed"
